Move main menu to its own component. Create menuContents component. Clean menu component from unused code.
