+++ src/components/menuContents.svelte
@@ -1,0 +1,27 @@
+<script lang="ts">
+  export let id: string;
+  export let visible = false;
+</script>
+
+<div {id} class="menu-contents {visible ? "visible": ""}">
+  <slot/>
+</div>
+
+<style>
+  .menu-contents {
+    box-sizing: border-box;
+    padding: 1em;
+    position: absolute;
+    top: 0;
+    left: calc(-1 * var(--mobile-width));
+    width: var(--mobile-width);
+    display: flex;
+    flex-direction: column;
+    overflow-y: hidden;
+    transition: left var(--menu-transition-time) var(--menu-transition-page-curve);
+  }
+
+  .menu-contents.visible {
+    left: 0;
+  }
+</style>
